pyproject.toml
[project]
name = "gesture"
version = "0.0.1"
description = "A series of classical computer vision algorithms for sign-language recognition."
authors = [{ name = "lavaman131", email = "lavaalex3@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.26.4",
    "pandas>=2.2.0",
    "opencv-python>=4.9.0.80",
    "fire>=0.5.0",
    "colorama>=0.4.6",
    "pyarrow>=15.0.0",
    "scikit-learn>=1.4.0",
    "matplotlib>=3.8.2",
    "seaborn>=0.13.2",
    "hydra-core>=1.3.2",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
